Refactor file structure to accommodate a single package.json. Combined package.jsons, eslintrc.jsons, and .gitignores.
