+++ package.json
@@ -4,13 +4,17 @@
   "description": "Repository for the Founders Program project \"Howl\". May 2019",
   "main": "server.js",
   "scripts": {
-    "dev:server": "nodemon ./src/server.js ",
-    "dev:client": "cd ./client && npm start",
-    "start": "node ./src/server.js",
-    "eslint": "eslint ./src",
-    "eslint:fix": "eslint --fix ./src",
+    "start": "react-scripts start",
+    "build": "SKIP_PREFLIGHT_CHECK=true react-scripts build",
+    "test": "react-scripts test",
+    "eject": "react-scripts eject",
+    "dev:start": "nodemon ./server/server.js ",
+    "eslint:src": "eslint ./src",
+    "eslint:fixsrc": "eslint --fix ./src",
+    "eslint:server": "eslint ./server",
+    "eslint:fixserver": "eslint --fix ./server",
     "coverage": "jest --coverage",
-    "heroku-postbuild": "cd client && npm i && npm run build"
+    "heroku-postbuild": "npm i && npm run build"
   },
   "repository": {
     "type": "git",
@@ -23,19 +27,51 @@
   },
   "homepage": "https://github.com/founders-programme-2/howl#readme",
   "dependencies": {
+    "@date-io/date-fns": "^1.3.5",
+    "@material-ui/core": "^4.0.0-beta.2",
+    "@material-ui/pickers": "^3.0.0-beta.1",
+    "babel-eslint": "^10.0.1",
+    "date-fns": "^2.0.0-alpha.27",
+    "env2": "^2.2.2",
+    "react": "^16.8.6",
+    "react-dom": "^16.8.6",
+    "react-router-dom": "^5.0.0",
+    "react-scripts": "3.0.1",
+    "styled-components": "^4.2.0",
+    "typescript": "^3.4.5",
     "airtable": "^0.6.0",
     "cloudinary": "^1.14.0",
     "datauri": "^2.0.0",
-    "env2": "^2.2.2",
     "express": "^4.16.4",
     "multer": "^1.4.1"
   },
   "devDependencies": {
+    "eslint-config-airbnb": "^17.1.0",
+    "eslint-config-prettier": "^4.1.0",
+    "eslint-plugin-jsx-a11y": "^6.2.1",
+    "eslint-plugin-prettier": "^3.0.1",
+    "eslint-plugin-react": "^7.12.4",
+    "prettier": "^1.17.0",
+    "react-a11y": "^1.1.0",
     "codecov": "^3.4.0",
     "eslint": "^5.16.0",
     "eslint-config-airbnb-base": "^13.1.0",
     "eslint-plugin-import": "^2.17.2",
-    "jest": "^24.8.0",
     "nodemon": "^1.19.0"
+  },
+  "eslintConfig": {
+    "extends": "react-app"
+  },
+  "browserslist": {
+    "production": [
+      ">0.2%",
+      "not dead",
+      "not op_mini all"
+    ],
+    "development": [
+      "last 1 chrome version",
+      "last 1 firefox version",
+      "last 1 safari version"
+    ]
   }
 }
